Add timeouts and don't pipe code to stdin.
We had to start using 'unbuffer' here because applications like python will buffer
their output if they don't think they are being ran in a tty. This allows us to get
output left over after we've destroyed the process instead of discarding it. To do
that, I'm using the 'unbuffer' program that comes with 'expect'. It works on many
operating systems but unfortinately isn't there by default in most of them. A second
option was to use the 'script' program, but it causes python (and I expect other
languages too) to return Windows-style newlines and carriage returns. If there is a
way to prevent that, 'script' would be the easier option because it would mean
nothing to install.